feat(babel-plugin): rpc worker pool and next app
