+++ @examples/nextjs/app/page.tsx
@@ -3,5 +3,5 @@ import {css} from 'taddy';
 console.log('check me');
 
 export default function Home() {
-    return <div {...css({color: 'orange'})}>hello boy</div>;
+    return <div {...css({color: 'purple'})}>hello boy!!!</div>;
 }
