+++ @taddy/babel-plugin/rollup.config.js
@@ -3,7 +3,7 @@ import commonConfig from '../../rollup.config.common';
 const config = {
     ...commonConfig,
 
-    input: ['index.ts'],
+    input: ['index.ts', './src/helpers/transformWorker.js'],
 };
 
 export default config;
